requirements.txt
# Content-Farm Bot – Dependencies
# Python 3.10+

# Content & Scripting (The Brain)
openai>=1.12.0
praw>=7.7.0

# Audio (The Voice)
elevenlabs>=1.0.0
edge-tts>=6.1.0

# Visuals (The Eyes)
httpx>=0.27.0
replicate>=0.25.0

# Video Engine (The Studio)
moviepy>=1.0.3
ffmpeg-python>=0.2.0
Pillow>=10.2.0

# Config & Env
python-dotenv>=1.0.0

# Optional: YouTube upload (API)
google-api-python-client>=2.115.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
